requirements.txt
paho-mqtt==1.3.1

requests==2.18.4